* commit 'c30ccfc1499a821ad1bdb2dfc0b4bd310a09a765':
  Tidy x86 assembler.

+++ compiler/dex/quick/mips/assemble_mips.cc
@@ -709,7 +709,7 @@ AssemblerStatus MipsMir2Lir::AssembleInstructions(CodeOffset start_addr) {
   return res;
 }
 
-int MipsMir2Lir::GetInsnSize(LIR* lir) {
+size_t MipsMir2Lir::GetInsnSize(LIR* lir) {
   DCHECK(!IsPseudoLirOp(lir->opcode));
   return EncodingMap[lir->opcode].size;
 }
